requirements.txt
streamlit
pandas
python-dotenv
google-generativeai>=0.7.0
pypdf
python-docx
